requirements.txt
asgiref==3.3.0
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
mysqlclient==1.3.4
psycopg2==2.8.6
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
